Rooによるコード整理
