+++ src/composables/usePerformanceOptimization.ts
@@ -10,75 +10,6 @@ import {
 } from "vue";
 
 /**
- * 高速なオブジェクトキャッシュ
- */
-export const useObjectCache = <T extends Record<string, unknown>>(
-  maxSize: number = 100
-) => {
-  const cache = markRaw(new Map<string, T>());
-
-  const set = (key: string, value: T): void => {
-    if (cache.size >= maxSize) {
-      // 最初のエントリを削除（FIFO）
-      const firstKey = cache.keys().next().value;
-      if (firstKey) {
-        cache.delete(firstKey);
-      }
-    }
-    cache.set(key, value);
-  };
-
-  const get = (key: string): T | undefined => cache.get(key);
-  const has = (key: string): boolean => cache.has(key);
-  const clear = (): void => cache.clear();
-  const size = (): number => cache.size;
-
-  return {
-    set,
-    get,
-    has,
-    clear,
-    size,
-  };
-};
-
-/**
- * 配列の差分検出とメモ化
- */
-export const useArrayMemoization = <T>(
-  accessor: () => readonly T[],
-  keyExtractor: (item: T) => string = (item) => String(item)
-) => {
-  const cachedArray = shallowRef<readonly T[]>([]);
-  const lastKeysHash = ref<string>("");
-
-  const memoizedArray = computed(() => {
-    const currentArray = accessor();
-    const currentKeys = currentArray.map(keyExtractor);
-    const currentHash = JSON.stringify(currentKeys);
-
-    // キーが変わっていない場合は既存の配列を返す
-    if (currentHash === lastKeysHash.value) {
-      return cachedArray.value;
-    }
-
-    lastKeysHash.value = currentHash;
-    cachedArray.value = currentArray;
-    return currentArray;
-  });
-
-  const forceUpdate = () => {
-    lastKeysHash.value = "";
-    triggerRef(cachedArray);
-  };
-
-  return {
-    memoizedArray,
-    forceUpdate,
-  };
-};
-
-/**
  * 計算プロパティの依存関係を最小化
  */
 export const useMinimalComputed = <T>(
