+++ src/utils/image.ts
@@ -213,6 +213,19 @@ export const getCardImageUrl = (cardId: string): Result<string, string> => {
 };
 
 /**
+ * カード画像URLを安全に取得
+ */
+export const getCardImageUrlSafe = (cardId: string): string => {
+  const result = getCardImageUrl(cardId);
+  if (result.isOk()) {
+    return result.value;
+  }
+  // エラーをログに記録
+  logger.warn(`Failed to get image URL for card: ${cardId}`, result.error);
+  return ""; // エラー時は空文字を返す
+};
+
+/**
  * 画像エラー時の処理
  */
 export const handleImageError = (event: Event): Result<void, string> => {
